renterd: move tracing out of internal
