+++ autopilot/migrator.go
@@ -6,7 +6,7 @@ import (
 	"sync"
 
 	"go.sia.tech/renterd/api"
-	"go.sia.tech/renterd/internal/tracing"
+	"go.sia.tech/renterd/tracing"
 	"go.uber.org/zap"
 )
